+++ support/chip-testing/test/cert-framework/tc-idm-4.1-support.test.ts
@@ -21,6 +21,7 @@ import { env } from "node:process";
 import type { SubscribeAndModifyTimeouts } from "../cert/tc-idm-4.1-support.js";
 import { subscribeAndModify } from "../cert/tc-idm-4.1-support.js";
 import { CertCheckFailedError } from "../cert/tc-support.js";
+import { fakeCertNode } from "./fake-cert-node.js";
 
 const SUBSCRIPTION_ID = 0x2a;
 
@@ -95,20 +96,7 @@ class Fixture {
     ) {
         this.#log = new LogFollower(this.#source, "th");
 
-        const unused = () => Promise.reject(new InternalError("not used by these tests"));
-        const node: CertNodeApi = {
-            invoke: unused,
-            invokeBatch: unused,
-            readAttribute: unused,
-            readAttributes: unused,
-            writeAttributes: unused,
-            readEvents: unused,
-            subscribeEvents: unused,
-            clientEndpoints: unused,
-            clientAttribute: unused,
-            openCommissioningWindow: unused,
-            operationalMdnsInstanceName: unused,
-            decommission: unused,
+        const node: CertNodeApi = fakeCertNode({
             subscribe: async (_path, opts) => {
                 this.#onUpdate = opts.onUpdate;
                 this.push(...subscribeRequestLines(PATH), ...subscribeResponseLines(SUBSCRIPTION_ID));
@@ -126,7 +114,7 @@ class Fixture {
             writeAttribute: async (_path, value) => {
                 this.onWrite(this, this.#writes++, value);
             },
-        };
+        });
 
         const device: CertDevice = {
             ...stubSubject(),